refactor(popup): index 파일 추출
