+++ packages/popup/src/index.ts
@@ -1,0 +1,2 @@
+export * from './popup'
+export { default } from './popup'
